Compromise for Ant 1.7
We wanted to deal with app names with spaces (SUITE-277), but we also want to work on CentOS 6 (SUITE-284) where we don't have the org.apache.tools.ant.taskdefs.optional.script package for the script task.  As a compromise, instead of bringing in another dependency, we'll try to make sure the app name doesn't have spaces before it gets to Ant.