Telling the user when the debug server starts.
This brings back more verbose output (that was lost with SUITE-283).  It's still not an ideal solution.  Until working in proper logging configuration, we take all the stdout from Ant and display it to the user.  This includes more than the user wants to see, but it does indicate when the debug server starts.  All stderr goes to the log file along with additional information to put errors in context (see SUITE-306).