Simplifying configuration of the outputTarget.
Actually this brings in a fix and an improvement. The fix is that the viewer's map now properly resizes not only when the toolbar is shown, but also when it is hidden and after it is laid out. The improvement is that the logic for ".tbar", ".bbar" etc. to point to a toolbar is now also used for the outputTarget, and not just for the actionTarget. The tutorial was updated to use the simpler configuration syntax, but also the previously suggested configuration works now. Fixes #SUITE-309.